// src/shared/components/form/form.types.ts
export interface Option {
  value: string;
  label: string;
}
// ...
export type RadioOption = {
  label: string;
  value: string;
  description?: string;
};
export interface ValidationRules {
  required?: string | boolean;
  minLength?: {
    value: number;
    message: string;
  };
  maxLength?: {
    value: number;
    message: string;
  };
  pattern?: {
    value: RegExp;
    message: string;
  };
  validate?: (value: string) => boolean | string;
}
export type FormElement =
  | {
    type: "radio";
    name: string;
    hasIcon?: boolean;
    label: string;
    options: RadioOption[];
    value: string;
    validation?: ValidationRules;
    onChange: (value: string) => void;
    notRequired?: boolean;
    colSpan?: number;
    disabled?: boolean;
    condition?: boolean;
  }
  | {
    type: "input";
    name: string;
    label: string;
    value: string;
    onChange: (value: string) => void;
    inputType: string;
    onBlur?: () => void;
    placeholder?: string;
    validation?: ValidationRules;
    min?: number;
    notRequired?: boolean;
    colSpan?: number;
    maxLength?: number;
    isLoading?: boolean;
    disabled?: boolean;
    condition?: boolean;
    numberOnly?: boolean;
  }
  | {
    type: "autocomplete";
    label: string;
    options: Option[];
    value: any;
    name: string;
    onChange: (value: any) => void;
    validation?: ValidationRules;
    notRequired?: boolean;
    colSpan?: number;
    isLoading?: boolean;
    disabled?: boolean;
    condition?: boolean;
    onClear?: () => void;
    autoSelectValue?: string | { value: string; label: string };

  }
  | {
    type: "checkbox";
    label: string;
    name: string;
    checked: boolean;
    validation?: ValidationRules;
    onChange: (value: boolean) => void;
    notRequired?: boolean;
    colSpan?: number;
    disabled?: boolean;
    className?: string;
    rules?: any;
    condition?: boolean;
  }
  | {
    type: "dateOfBirth";
    hijriLabel: string;
    gregorianLabel: string;
    notRequired?: boolean;
    colSpan?: number;
    hijriFieldName?: string;
    gregorianFieldName?: string;
    name?: string;
    value?: any;
    showDateGregorian?: any;
    validation?: any;
    rules?: any;
    condition?: boolean;
  }
  | {
    type: "readonly";
    label: string;
    value: string;
    notRequired?: boolean;
    colSpan?: number;
    isLoading?: boolean;
    condition?: boolean;
  }
  | {
    type: "custom";
    component: JSX.Element;
    colSpan?: number;
    condition?: boolean;
    name?: string;
  }
  | {
    type: "button";
    label: string;
    onClick: () => void;
    name?: string;
    size?: "xs" | "xs20" | "sm" | "md" | "sl" | "lg";
    className?: string;
    colSpan?: number;
    disabled?: boolean;
    condition?: boolean;
  };